+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 4.6.0
 * Changed all occurrences of ``Integer`` and ``Double`` to ``Number`` for improved compatibility with Leaflet's API
+* Updated OpenStreetMap's Tile-Server address to ``tile.openstreetmap.org`` [openstreetmap/operations#737](https://github.com/openstreetmap/operations/issues/737)
 
 ## 4.5.0
 * Added support for [WMS](https://leafletjs.com/reference.html#tilelayer-wms) #486